refactor: move ComputePipelineModal sub-components to separate files
