+++ packages/core/components/Modal/ComputePipelineModal/index.tsx
@@ -10,11 +10,13 @@ import AnnotationName from "../../../entity/Annotation/AnnotationName";
 import { Pipeline, PipelineParameter } from "../../../entity/ComputePipeline";
 import FileSelection from "../../../entity/FileSelection";
 import { interaction, selection } from "../../../state";
+import PipelineInfoCard from "./PipelineInfoCard";
+import PipelineParameterInput from "./PipelineParameterInput";
+import SelectedFilesExpander from "./SelectedFilesExpander";
 
 import styles from "./ComputePipelineModal.module.css";
 
 const FILE_PATHS_PARAM = "file_paths";
-const MAX_VISIBLE_FILE_PATHS = 10;
 
 type ModalPhase = "loading" | "selecting" | "configuring" | "submitting" | "submitted" | "error";
 
@@ -35,123 +37,6 @@ function validateParam(param: PipelineParameter, value: unknown): string {
     if (validation.max !== undefined && n > validation.max)
         return `Must be no more than ${validation.max}.`;
     return "";
-}
-
-interface ParamInputProps {
-    param: PipelineParameter;
-    value: unknown;
-    error: string;
-    onChange: (value: unknown) => void;
-    onBlur: () => void;
-}
-
-function PipelineParameterInput({ param, value, error, onChange, onBlur }: ParamInputProps) {
-    return (
-        <div className={styles.paramField}>
-            <div className={styles.paramLabel}>
-                {param.label}
-                {param.required && " *"}
-            </div>
-            <div className={styles.paramDesc}>{param.description}</div>
-            <TextField
-                type="number"
-                value={value !== null && value !== undefined ? String(value) : ""}
-                onChange={(_, v) => onChange(v ?? "")}
-                onBlur={onBlur}
-                placeholder={param.default !== null ? String(param.default) : ""}
-                borderless
-                styles={{
-                    field: {
-                        backgroundColor: "var(--secondary-background-color)",
-                        color: "var(--secondary-text-color)",
-                    },
-                }}
-            />
-            {error && <div className={styles.paramError}>{error}</div>}
-        </div>
-    );
-}
-
-interface SelectedFilesExpanderProps {
-    count: number;
-    onExpand: () => Promise<string[]>;
-}
-
-function SelectedFilesExpander({ count, onExpand }: SelectedFilesExpanderProps) {
-    const [expanded, setExpanded] = React.useState(false);
-    const [paths, setPaths] = React.useState<string[] | null>(null);
-    const [loading, setLoading] = React.useState(false);
-
-    const toggle = async () => {
-        if (!expanded && paths === null) {
-            setLoading(true);
-            const fetched = await onExpand();
-            setPaths(fetched);
-            setLoading(false);
-        }
-        setExpanded((e) => !e);
-    };
-
-    return (
-        <div className={styles.expander}>
-            <div className={styles.expanderHeader} onClick={toggle}>
-                <span className={styles.expanderHeaderLeft}>
-                    <Icon iconName="FolderOpen" />
-                    {count} selected file{count !== 1 ? "s" : ""}
-                </span>
-                {loading ? <Spinner /> : <Icon iconName={expanded ? "ChevronUp" : "ChevronDown"} />}
-            </div>
-            {expanded && paths !== null && (
-                <div className={styles.expanderBody}>
-                    <div className={styles.expanderNote}>
-                        Files are added automatically from your current selection.
-                    </div>
-                    <ul className={styles.filePathList}>
-                        {paths.slice(0, MAX_VISIBLE_FILE_PATHS).map((p) => (
-                            <li key={p} className={styles.filePathItem}>
-                                <Icon iconName="TextDocument" />
-                                {p}
-                            </li>
-                        ))}
-                    </ul>
-                    {paths.length > MAX_VISIBLE_FILE_PATHS && (
-                        <div className={styles.expanderNote}>
-                            …and {paths.length - MAX_VISIBLE_FILE_PATHS} more
-                        </div>
-                    )}
-                </div>
-            )}
-        </div>
-    );
-}
-
-interface PipelineInfoCardProps {
-    pipeline: Pipeline;
-}
-
-function PipelineInfoCard({ pipeline }: PipelineInfoCardProps) {
-    return (
-        <div className={styles.infoCard}>
-            <div className={styles.infoCardDesc}>{pipeline.description}</div>
-            <hr className={styles.infoCardDivider} />
-            <div className={styles.infoTable}>
-                <div className={styles.infoRow}>
-                    <span className={styles.infoRowLabel}>Accepted formats</span>
-                    <span className={styles.infoRowValue}>
-                        {pipeline.acceptedExtensions.map((e) => e.toUpperCase()).join(", ")}
-                    </span>
-                </div>
-                <div className={styles.infoRow}>
-                    <span className={styles.infoRowLabel}>Run locations</span>
-                    <span className={styles.infoRowValue}>{pipeline.clusters.join(", ")}</span>
-                </div>
-                <div className={styles.infoRow}>
-                    <span className={styles.infoRowLabel}>Restrictions</span>
-                    <span className={styles.infoRowValue}>{pipeline.restrictions ?? "None"}</span>
-                </div>
-            </div>
-        </div>
-    );
 }
 
 export default function ComputePipelineModal({ onDismiss }: ModalProps) {
@@ -508,12 +393,7 @@ export default function ComputePipelineModal({ onDismiss }: ModalProps) {
                                 }}
                                 placeholder="Enter your user ID (ex. first.last)"
                                 borderless
-                                styles={{
-                                    field: {
-                                        backgroundColor: "var(--secondary-background-color)",
-                                        color: "var(--secondary-text-color)",
-                                    },
-                                }}
+                                className={styles.textField}
                             />
                             {userIdError && <div className={styles.paramError}>{userIdError}</div>}
                         </div>
